Implemented the constructor and destructor for the two arrays that will house the values of ships (once the ship placement functions are implemented).
Both grids get initialized to the value "0". In the future if a ship is placed
on the grid, the value at that coordinate will be overridden to a different
number, such as "2".

Player One's grid and Player Two's grid are both outputted into a visually
appealing board, where the rows and columns correspond to the units on the board.