Created an implementation file for the ship class. The ship class is responsible for knowing the length, direction, and start of the ship.
isDestroyed function checks the integrity of the ship, and once the ship
integrity is at 0, then the ship is considered destroyed.

Need to implement a hit function, that takes a valid string from the user,
which represents the column and row (i.e A1-J10). Applying the hit function to
another player would hopefully update the player's grid, and show whether
the ship was hit or not.