The implementation file for the Player class.
Implemented a constructor, which reads the amount of ships that are going
to be in use, and fills the 2D char array with "O" at every index.

Implemented a showWaters function, which currently just outputs the 2D char
array in order to confirm that the values of the indexes were initialized to "O".

The anchorShips function still needs implementation. The goal of the function
is for the user to place all of their ships at one time.
(i.e. the user has 4 ships, so the user needs to place down 4 ships based
on starting location, orientation, and length)
--Current issue is validating correct user input.